Oppgrader felles vedtaksid til int (#469)
* Oppgrader felles vedtaksid til int

* Oppgrader felles vedtaksid til int

+++ src/test/kotlin/no/nav/bidrag/vedtak/TestUtil.kt
@@ -1056,7 +1056,7 @@ class TestUtil {
             behandlingsreferanseListe = byggBehandlingsreferanseListe(),
         )
 
-        fun byggVedtaksforslagMedOppdatertInnholdRequest(sisteVedtaksid: Long?) = OpprettVedtakRequestDto(
+        fun byggVedtaksforslagMedOppdatertInnholdRequest(sisteVedtaksid: Int?) = OpprettVedtakRequestDto(
             kilde = Vedtakskilde.AUTOMATISK,
             type = Vedtakstype.ALDERSJUSTERING,
             opprettetAv = "X654321",
